Major revision; can install git/zip urls
 - changed API for existing libs
 - optionally cn create backups

+++ lib/arli/installers/zip_file.rb
@@ -2,36 +2,44 @@ require 'archive/zip'
 
 module Arli
   module Installers
-    class Unzipper
-      class UnzipperError < StandardError;
-      end
-
+    class ZipFile
       attr_accessor :lib, :lib_dir
 
       def initialize(lib:)
-        self.lib = lib
+        self.lib     = lib
         self.lib_dir = lib.name.gsub(/ /, '_')
         raise 'Invalid URL for this installer: ' + lib.url unless lib.url =~ /\.zip$/i
       end
 
       def install
         download!
-        dirs = dirs_matching(lib_dir)
-        dirs.delete(lib_dir)
 
-        # Delete old ones if any
-        dirs.each { |d| FileUtils.rm_f(d) }
+        remove_library!
+        remove_library_versions!
 
         unzip(zip_archive, '.')
+
         dir = dirs_matching(lib_dir).first
-        FileUtils.move(dir, lib.name) if dir && dir != lib_dir
-        FileUtils.rm_f(zip_archive)
+
+        FileUtils.move(dir, lib_dir) if dir
+
+        FileUtils.rm_f(zip_archive) if File.exist?(zip_archive)
+      end
+
+      def remove_library!
+        FileUtils.rm_rf(lib_dir)
+      end
+
+      def remove_library_versions!
+        dirs = dirs_matching(lib_dir)
+        dirs.delete(lib_dir) # we don't want to delete the actual library
+        dirs.each { |d| FileUtils.rm_f(d) }
       end
 
       private
 
       def dirs_matching(name)
-        Dir.glob("#{name}*").select { |d| File.directory?(d) }
+        Dir.glob("#{name}*").select { |d| File.directory?(d)  }
       end
 
       def zip_archive
@@ -46,7 +54,7 @@ module Arli
       #   Archive::Zip.extract(file, destination, on_error: :skip)
       # end
       def unzip(file, destination)
-        `unzip #{file} -d #{destination}`
+        `unzip -o #{file} -d #{destination}`
       end
     end
   end
